Rename test sector1 to sector01
